Support source maps in resplendence
A thing I learned about webpack source maps is that apparently they have no concept of a "copied character range", meaning that even if a span of characters from the source is 1:1 identical to the same span in the generated code, you will have to manually map each symbol to itself. One way to do this is character-by-character, but in an effort to not drag the performance down too much I instead split everything by clusters of alphaNumberic chars and skipped over whitespace and symbol.

This change severely impacts resplendence's performance, and I'm not sure of a good way to work around it. I managed to tune it so that it's not unliveably slow, but I think another look at it will be warranted soon. It might be the use of regex that is dragging things down, it's hard to say.

Some other changes that happened along with this: tests and jsdoc typings. I've started to become real fan of skipping typescript and just having the TS compiler do the checks using the documentation. This library doesn't run through babel or any other code transforms and it feels nice to know that the code you write is the code that is getting run.