Don't break source maps
So I went through a heroic effort to learn and understand source maps and to generate proper source maps for resplendence, and a few days after finishing it all I realized that it would be both simpler and more useable to simply tweak the code generation a bit so that we won't need them.

This PR modifies the script-loader and style loader to preserve the correct line numbers in the generated source. The script loader does this by turning the template strings into comments (I really like how this ends up looking!) The style loader just makes sure that there are a bunch of empty newlines where the js code used to be, because the comment system ended up too cluttered.

Some other changes that happened along with this: tests and jsdoc typings. I've started to become real fan of skipping typescript and just having the TS compiler do the checks using the documentation. This library doesn't run through babel or any other code transforms and it feels nice to know that the code you write is the code that is getting run.